test: split code to conftest.py
