+++ tests/integration/conftest.py
@@ -11,3 +11,13 @@ def pixi() -> Path:
 @pytest.fixture
 def test_data() -> Path:
     return Path(__file__).parent.joinpath("test_data").resolve()
+
+
+@pytest.fixture
+def dummy_channel_1(test_data: Path) -> str:
+    return test_data.joinpath("dummy_channel_1/output").as_uri()
+
+
+@pytest.fixture
+def dummy_channel_2(test_data: Path) -> str:
+    return test_data.joinpath("dummy_channel_2/output").as_uri()
